Split the renderer
