+++ bevy_kayak_ui/src/render/font/extract.rs
@@ -4,12 +4,12 @@ use bevy::{
     sprite::Rect,
 };
 use kayak_core::render_primitive::RenderPrimitive;
-use kayak_core::styles::Corner;
 use kayak_font::{Alignment, CoordinateSystem, KayakFont};
 
-use crate::{
+use crate::to_bevy_color;
+use bevy_kayak_renderer::{
     render::unified::pipeline::{ExtractQuadBundle, ExtractedQuad, UIQuadType},
-    to_bevy_color,
+    Corner,
 };
 
 use super::font_mapping::FontMapping;
